Rename file. Refactor ExtensionMessenger.
